backend/requirements.txt
Django
djangorestframework
python-dotenv
openai
django-cors-headers